Extract tmux resume parser
